requirements/base.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==3.1.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.1.0
autobahn==19.5.1
Automat==0.7.0
billiard==3.5.0.5
boto3==1.9.137
botocore==1.12.137
celery==4.2.1
certifi==2019.3.9
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
click==6.7
constantly==15.1.0
daphne==2.3.0
dj-database-url==0.5.0
Django==2.0.9
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-celery-beat==1.2.0
django-crispy-forms==1.7.2
django-filter==2.1.0
django-fsm==2.6.1
django-model-utils==3.1.2
django-qr-code==1.0.0
django-storages==1.7.1
django-tables2==2.0.6
django-timezone-field==3.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docutils==0.14
drf-nested-routers==0.91
EasyProcess==0.2.7
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
jmespath==0.9.4
kombu==4.5.0
lxml==4.3.3
Markdown==3.1
msgpack==0.6.1
pep8==1.7.1
Pillow==6.0.0
psycopg2-binary==2.7.7
pyaxmlparser==0.3.15
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
PySocks==1.6.8
python-crontab==2.3.6
python-dateutil==2.8.0
pytz==2019.1
PyVirtualDisplay==0.2.3
qrcode==6.1
redis==3.2.1
requests==2.21.0
s3transfer==0.2.0
selenium==3.141.0
six==1.12.0
sorl-thumbnail==12.5.0
spotipy==2.4.4
twilio==6.26.2
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.2
vine==1.3.0
websockets==7.0
zope.interface==4.6.0
